requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cssselect2==0.8.0
Django==5.1.4
django-appconf==1.1.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-redis==5.4.0
django-tinymce==4.1.0
fonttools==4.57.0
frozenlist==1.5.0
idna==3.10
kombu==5.5.3
multidict==6.2.0
phonenumbers==8.13.54
pillow==11.1.0
prompt_toolkit==3.0.51
propcache==0.3.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
rcssmin==1.1.2
redis==6.0.0
requests==2.32.3
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.5.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
